moving docs to mkdocs (#856)
* moving docs to mkdocs

* transfering readme to md and more

* fixing build

* splitting usage.md

* improving schema.md and index.md

* fix make_history.rst

* models intro

* working on data conversation and required fields

* more fixes to models.md

* list all standard types supported

* list of pydantic types

* tweaks

* update links in code

* Apply suggestions from code review

incorporate @dmontagu's suggestions.


* Apply suggestions from code review

more missed suggestions.


* Apply suggestions from code review

more corrects.

* cleanup

* Field order warning

* fix and regenerate benchmarks

* format examples better, cleanup

* improve schema mapping table

* correct highlighting file types in schema.md

* add redirects in javascript

* add logo

+++ changes/make_history.py
@@ -6,20 +6,22 @@ from importlib.machinery import SourceFileLoader
 from pathlib import Path
 
 THIS_DIR = Path(__file__).parent
-name_regex = re.compile(r'(\d+)-(.*?)\.rst')
+name_regex = re.compile(r'(\d+)-(.*?)\.md')
 bullet_list = []
 
-for p in THIS_DIR.glob('*.rst'):
+for p in THIS_DIR.glob('*.md'):
+    if p.name == 'README.md':
+        continue
     m = name_regex.fullmatch(p.name)
     if not m:
         raise RuntimeError(f'{p.name!r}: invalid change file name')
     gh_id, creator = m.groups()
-    content = p.read_text().strip('\n. ').replace('\r\n', '\n')
+    content = p.read_text().replace('\r\n', '\n').strip('\n. ')
     if '\n\n' in content:
         raise RuntimeError(f'{p.name!r}: content includes multiple paragraphs')
     content = content.replace('\n', '\n  ')
-    order = 0 if '**breaking change' in content.lower() else 1
-    bullet_list.append((order, int(gh_id), f'* {content}, #{gh_id} by @{creator}'))
+    priority = 0 if '**breaking change' in content.lower() else 1
+    bullet_list.append((priority, int(gh_id), f'* {content}, #{gh_id} by @{creator}'))
 
 if not bullet_list:
     print('no changes found')
@@ -27,22 +29,18 @@ if not bullet_list:
 
 version = SourceFileLoader('version', 'pydantic/version.py').load_module()
 chunk_title = f'v{version.VERSION} ({date.today():%Y-%m-%d})'
-bullets = '\n'.join(c for *_, c in sorted(bullet_list))
-new_chunk = f"""
-{chunk_title}
-{'.' * len(chunk_title)}
-{bullets}
-"""
+new_chunk = '## {}\n{}\n'.format(chunk_title, '\n'.join(c for *_, c in sorted(bullet_list)))
 
 print(f'{chunk_title}...{len(bullet_list)} items')
-history_path = THIS_DIR / '..' / 'HISTORY.rst'
+history_path = THIS_DIR / '..' / 'HISTORY.md'
 history = history_path.read_text()
-history = re.sub(r'(History\n-------\n)', fr'\1{new_chunk}', history, count=1)
+
 history_path.write_text(history)
-for p in THIS_DIR.glob('*.rst'):
-    p.unlink()
+for p in THIS_DIR.glob('*.md'):
+    if p.name != 'README.md':
+        p.unlink()
 
 print(
-    'changes deleted and HISTORY.rst successfully updated, to reset use:\n\n'
-    '  git checkout -- changes/*.rst HISTORY.rst\n'
+    'changes deleted and HISTORY.md successfully updated, to reset use:\n\n'
+    '  git checkout -- changes/*-*.md HISTORY.md\n'
 )
